pyproject.toml
[project]
name = "ai-agent-own-written-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.13.0",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.2",
    "wikipedia>=1.4.0",
]
